enable 32-bit build support

+++ prov/rxd/src/rxd_av.c
@@ -108,12 +108,12 @@ close:
 
 static fi_addr_t rxd_av_dg_addr(struct rxd_av *av, fi_addr_t fi_addr)
 {
-	fi_addr_t rxd_addr = (fi_addr_t) ofi_idx_lookup(&av->fi_addr_idx,
+	fi_addr_t rxd_addr = (fi_addr_t)(uintptr_t) ofi_idx_lookup(&av->fi_addr_idx,
 					     RXD_IDX_OFFSET((int)fi_addr));
 	if (!rxd_addr)
 		return FI_ADDR_UNSPEC;
 
-	return (fi_addr_t) ofi_idx_lookup(&av->rxdaddr_dg_idx, (int)rxd_addr);
+	return (fi_addr_t)(uintptr_t) ofi_idx_lookup(&av->rxdaddr_dg_idx, (int)rxd_addr);
 }
 
 static int rxd_set_rxd_addr(struct rxd_av *av, fi_addr_t dg_addr, fi_addr_t *addr)
@@ -169,7 +169,7 @@ int rxd_av_insert_dg_addr(struct rxd_av *av, const void *addr,
 		goto nomem;
 	}
 
-	ret = ofi_rbmap_insert(&av->rbmap, (void *)addr, (void *)(*rxd_addr),
+	ret = ofi_rbmap_insert(&av->rbmap, (void *)addr, (void *)(uintptr_t)(*rxd_addr),
 			       NULL);
 	if (ret) {
 		assert(ret != -FI_EALREADY);
@@ -213,7 +213,7 @@ static int rxd_av_insert(struct fid_av *av_fid, const void *addr, size_t count,
 	for (; i < count; i++, addr = (uint8_t *) addr + av->dg_addrlen) {
 		node = ofi_rbmap_find(&av->rbmap, (void *) addr);
 		if (node) {
-			rxd_addr = (fi_addr_t) node->data;
+			rxd_addr = (fi_addr_t)(uintptr_t) node->data;
 		} else {
 			ret = rxd_av_insert_dg_addr(av, addr, &rxd_addr,
 						    flags, sync_err ?
@@ -352,7 +352,7 @@ static int rxd_av_close(struct fid *fid)
 		return ret;
 
 	while ((node = ofi_rbmap_get_root(&av->rbmap))) {
-		rxd_addr = (fi_addr_t) node->data;
+		rxd_addr = (fi_addr_t)(uintptr_t) node->data;
 		dg_addr = (intptr_t)ofi_idx_lookup(&av->rxdaddr_dg_idx,
 						   (int) rxd_addr);
